# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_JobView_": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/JobView"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[JobView]",
        "type": "object"
      },
      "ApiResponse_OntologySummaryView_": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/OntologySummaryView"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[OntologySummaryView]",
        "type": "object"
      },
...
      "ApiResponse_tuple_JobView__________": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "title": "Data",
            "type": "array"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[tuple[JobView, ...]]",
        "type": "object"
      },
      "ApiResponse_tuple_RunView__________": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "title": "Data",
            "type": "array"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[tuple[RunView, ...]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StageStatusView": {
        "description": "Whether the pipeline stage behind a payload actually produced it.\n\n`NOT_RUNNABLE` is what an endpoint returns when its owning module has not been built --\nmodules 7, 8 and 15 today. It is a 200 with a full envelope and an explicit statement,\nnever an empty success: `docs/architecture.md` \u00a72 requires that a partially computed\nrun \"returns what exists with the stage stated, never a silently truncated graph\", and\nan empty list with no explanation is exactly the silent truncation that forbids.",
        "enum": [
          "COMPLETE",
          "PARTIAL",
          "NOT_RUNNABLE"
        ],
        "title": "StageStatusView",
        "type": "string"
      },
      "StartJobRequest": {
        "additionalProperties": false,
        "description": "What a caller supplies to start one pipeline execution.",
        "properties": {
          "dataset_id": {
            "minLength": 1,
            "title": "Dataset Id",
            "type": "string"
          },
          "rows": {
            "default": 150,
            "minimum": 0.0,
            "title": "Rows",
            "type": "integer"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "dataset_id"
        ],
        "title": "StartJobRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A domain-agnostic causal intelligence engine. Every response carries its run, its provenance, its evidence references and its confidence decomposition; no endpoint returns a confidence without its components.",
    "title": "CausaLog",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "description": "List recent executions, newest first.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_tuple_JobView__________"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Accept a pipeline execution and return 202 with its identifier.\n\n202, not 201: nothing is created yet that the caller can read in full. The execution\nexists and has an identifier; its artifacts do not.\n\nA repeated `Idempotency-Key` returns the ORIGINAL execution rather than starting a\nsecond one, and the lookup happens before anything is written.",
        "operationId": "start_job_v1_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JobView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{execution_id}": {
      "get": {
        "description": "Return one execution's stage-by-stage progress and timing.",
        "operationId": "read_job_v1_jobs__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JobView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/runs": {
      "get": {
        "description": "List every run materialized in this process, in canonical sequence.\n\nCATALOG scope: this body spans many runs, so there is no single `OutputEnvelope` that\ndescribes it. Each ITEM carries the five inputs that identify its run; synthesizing an\nenvelope for the listing would be fabricating version fields that describe nothing.",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_tuple_RunView__________"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}/ontology": {
      "get": {
        "description": "Return the ontology this run reasoned under, with its hash.",
        "operationId": "read_ontology_v1_runs__run_id__ontology_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_OntologySummaryView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Ontology",
        "tags": [
          "introspection"
        ]
      }
    },
...
    "/v1/runs/{run_id}/rule-pack": {
      "get": {
        "description": "Return the rules that governed this run's conclusions, and their coverage gaps.",
        "operationId": "read_rule_pack_v1_runs__run_id__rule_pack_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RulePackView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Rule Pack",
        "tags": [
          "introspection"
        ]
      }
    },
...
  }
}